:recycle: Move color param transform to callback
